refactor(EncryptionContactChip): Put EncryptionStatus in its own file
